skills/ai-daily-news/requirements.txt
# AI日报技能依赖
feedparser>=6.0.0
requests>=2.28.0
PyYAML>=6.0
beautifulsoup4>=4.12.0
